Rename worker and export processBlueprint
